+++ include/pico/codegen/backend-direct/polymorphic.h
@@ -4,7 +4,7 @@
 #include "pico/binding/address_env.h"
 #include "pico/syntax/syntax.h"
 #include "pico/codegen/codegen.h"
-#include "pico/codegen/internal.h"
+#include "pico/codegen/backend-direct/internal.h"
 
 void generate_polymorphic(SymbolArray arr, Syntax syn, AddressEnv* env, Target target, InternalLinkData* links, Allocator* a, ErrorPoint* point);
