+++ include/pico/codegen/foreign_adapters.h
@@ -5,9 +5,5 @@
 #include "pico/values/types.h"
 #include "pico/values/ctypes.h"
 
-bool can_convert(CType* ctype, PiType* ptype); 
-void convert_c_fn(void* cfn, CType* ctype, PiType* ptype, Assembler* ass, Allocator* a, ErrorPoint* point); 
-
-bool can_reinterpret(CType* ctype, PiType* ptype); 
 
 #endif
