+++ src/pico/analysis/typecheck.c
@@ -10,7 +10,6 @@
 #include "pico/analysis/typecheck.h"
 #include "pico/values/ctypes.h"
 #include "pico/codegen/codegen.h"
-#include "pico/codegen/foreign_adapters.h"
 #include "pico/eval/call.h"
 #include "pico/stdlib/core.h"
 #include "pico/stdlib/meta/meta.h"
