+++ src/pico/stdlib/extra.c
@@ -11,8 +11,8 @@
 #include "pico/stdlib/meta/meta.h"
 #include "pico/syntax/concrete.h"
 #include "pico/analysis/abstraction.h"
-#include "pico/codegen/foreign_adapters.h"
-#include "pico/codegen/internal.h"
+#include "pico/codegen/codegen.h"
+#include "pico/codegen/backend-direct/internal.h"
 
 
 static jump_buf* m_buf;
