+++ src/pico/stdlib/meta/refl.c
@@ -3,7 +3,7 @@
 
 #include "pico/values/modular.h"
 #include "pico/values/types.h"
-#include "pico/codegen/foreign_adapters.h"
+#include "pico/codegen/codegen.h"
 #include "pico/stdlib/core.h"
 #include "pico/stdlib/extra.h"
 #include "pico/stdlib/platform/submodules.h"
