+++ src/pico/stdlib/platform/hedron.c
@@ -2,7 +2,7 @@
 #include "platform/hedron/hedron.h"
 
 #include "pico/values/ctypes.h"
-#include "pico/codegen/foreign_adapters.h"
+#include "pico/codegen/codegen.h"
 #include "pico/stdlib/core.h"
 #include "pico/stdlib/platform/submodules.h"
