+++ src/pico/stdlib/platform/window.c
@@ -2,7 +2,7 @@
 #include "platform/window/window.h"
 
 #include "pico/values/ctypes.h"
-#include "pico/codegen/foreign_adapters.h"
+#include "pico/codegen/codegen.h"
 #include "pico/stdlib/core.h"
 #include "pico/stdlib/extra.h"
 #include "pico/stdlib/platform/submodules.h"
